feat: gamification, events, guides, audio tours, seed data expansion, and i18n overhaul
- Add Badge, UserBadge, Guide, GuidePlace, Event, EventRegistration, AudioGuide models with migrations
- Add API endpoints: badges, events, guides, audio_guides, community, life, reviews
- Expand seed data: 58 places, 31 activities, 10 badges, 6 events, 4 guides
- Add feedback table and community features
- Fix conftest.py to import all models for SQLite test DB
- Fix Badge table UUID PK (drop+recreate migration)
- Add StatusBar labels for badges/events/guides (remove '(models pending)')
- Expand i18n: complete ar/en/fr locale files with all new pages
- Add new pages: Explore, Meetups, Mood, PhotoWalk, Tourist, Wellness
- Update proxy.ts for Next.js 16 middleware convention
- Clean up dependencies, fix tsconfig, update manifest
- Fix blank page on production build (zombie next-server process)

+++ backend/app/models/auth_token.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from datetime import datetime
-from sqlalchemy import String, ForeignKey, DateTime, Boolean, BigInteger
+from sqlalchemy import String, ForeignKey, DateTime, Boolean, Integer
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.db.base_class import Base
@@ -9,7 +9,7 @@ from app.db.base_class import Base
 class RefreshToken(Base):
     __tablename__ = "refresh_tokens"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, index=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, index=True)
     token_hash: Mapped[str] = mapped_column(String(255), unique=True, index=True, nullable=False)
     user_id: Mapped[int] = mapped_column(ForeignKey("users.id", ondelete="CASCADE"), nullable=False)
     expires_at: Mapped[datetime] = mapped_column(DateTime, nullable=False)
